Added Slack API call to verify user admin status
Complete task 1, verifying that the requesting user_id is a user who is a Slack admin. This ostensibly serves to limit the pool of members of the Slack team who can just add a Github user.

In my gut I know this measure is not enough for real security. If this were deployed as is, the deployer would have to assume anyone with access to the Slack team has the ability to utilize the provided Github authentication token.  To spell it out, they could observe the network traffic when issuing a /slackager slash command, view the "token" parameter, and then obtain an admin's user id by calling using the [users.list](https://api.slack.com/methods/users.list) Web API call by creating their own API token. After that spoofing would be a matter of fililng in the admin user's id in place of their own.

A whitelist of "okay"ed users would face the same challenge. Ultimately the deployer probably needs an additional secret when running `/slackager` to authenticate. This seems like it would make usability a challenge - but also like table stakes for any true sense of security.

Commiting this with security flaws because it moves the ball forward, but will add an issue about the security and revisit later.